% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars.R
\name{plot_bars}
\alias{plot_bars}
\title{Produire un diagramme en barres simple}
\usage{
plot_bars(.data, x)
}
\arguments{
\item{.data}{La base de données (tibble)}

\item{x}{La variable étudiée}
}
\description{
Produire un diagramme en barres simple
}
\examples{
covid7 \%>\% plot_bars(pcs)

}
