% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleveland.R
\name{plot_cleveland}
\alias{plot_cleveland}
\title{Produire un diagramme de Cleveland simple}
\usage{
plot_cleveland(.data, x)
}
\arguments{
\item{.data}{La base de données (tibble)}

\item{x}{La variable étudiée}
}
\description{
Produire un diagramme de Cleveland simple
}
\examples{
covid7 \%>\% plot_cleveland(appareil)

}
