% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Produire un histogramme simple}
\usage{
plot_hist(.data, x)
}
\arguments{
\item{.data}{La base de données (tibble)}

\item{x}{La variable étudiée}
}
\description{
Produire un histogramme simple
}
\examples{
covid7 \%>\% plot_hist(logement_pieces)

}
