% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Produire un nuage de points simple}
\usage{
plot_points(.data, x)
}
\arguments{
\item{.data}{La base de données (tibble)}

\item{x}{La variable étudiée}
}
\description{
Produire un nuage de points simple
}
\examples{
covid7 \%>\% plot_points(logement_pieces)

}
