% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lesson.R
\name{run_lesson}
\alias{run_lesson}
\title{Lancer un tutoriel de "socstat"}
\usage{
run_lesson(lesson_name)
}
\arguments{
\item{lesson_name}{Le nom de la leçon}
}
\description{
Lancer un tutoriel de "socstat"
}
\examples{
run_lesson("intro")

}
