% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_univ.R
\name{table_univ}
\alias{table_univ}
\title{Produire un tri à plat simple sur une variable}
\usage{
table_univ(.data, x)
}
\arguments{
\item{.data}{La base de données (tibble)}

\item{x}{La variable étudiée}
}
\description{
Produire un tri à plat simple sur une variable
}
\examples{
covid7 \%>\% table_univ(pcs)

}
